requirements.txt
streamlit
websockets
python-dotenv
boto3